refactored the database tables and deleted the extraneous models and repositories
