+++ src/main/java/com/questing/littlequest/controllers/StoryController.java
@@ -1,8 +1,8 @@
 package com.questing.littlequest.controllers;
 
 import com.questing.littlequest.models.Stories;
-import com.questing.littlequest.repositories.ChoicesRepository;
-import com.questing.littlequest.repositories.StoriesRepository;
+import com.questing.littlequest.repositories.ChoiceRepository;
+import com.questing.littlequest.repositories.StoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -14,14 +14,14 @@ import java.util.List;
 public class StoryController {
 
     @Autowired
-    StoriesRepository storiesRepository;
+    StoryRepository storyRepository;
 
     @Autowired
-    ChoicesRepository choicesRepository;
+    ChoiceRepository choiceRepository;
 
     @GetMapping("/story-choice")
     public String storyDisplay(Model model) {
-        List<Stories> stories = storiesRepository.findAll();
+        List<Stories> stories = storyRepository.findAll();
         System.out.println("Stories = " + stories.toString());
         model.addAttribute("stories", stories);
         return "redirect:/story-choice";
