+++ src/main/java/com/questing/littlequest/models/Choices.java
@@ -1,29 +1,40 @@
 package com.questing.littlequest.models;
 
+
 import javax.persistence.*;
 
 @Entity
-@Table(name = "story_choices")
+    @Table(name = "choices")
 public class Choices {
     @Id
     @GeneratedValue
-    @SequenceGenerator(name="choices-generator")
+    @SequenceGenerator(name="choice-id-generator")
 
     public Long choice_id;
     public Long story_id;
-    public Long previous_choice_id;
+    public Long prompt_id;
     public String choice_a_text;
     public String choice_b_text;
-    public Long text_id;
+
+
+
+//    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "prompt_id", nullable = false)
+    private Prompts newPromptTable;
 
     public Choices() {
     }
 
-    public Choices(Long story_id, Long previous_choice_id, String choice_a_text, String choice_b_text, Long text_id) {
+    public Choices(Long story_id, Long prompt_id, String choice_a_text, String choice_b_text) {
         this.story_id = story_id;
-        this.previous_choice_id = previous_choice_id;
+        this.prompt_id = prompt_id;
         this.choice_a_text = choice_a_text;
         this.choice_b_text = choice_b_text;
-        this.text_id = text_id;
+    }
+
+    public String toString() {
+        String choices = "Choice A: " + this.choice_a_text + "\n" +
+                "Choice B: " + this.choice_b_text;
+        return choices;
     }
 }
