+++ src/main/java/com/questing/littlequest/models/Stories.java
@@ -2,28 +2,27 @@ package com.questing.littlequest.models;
 
 import javax.persistence.*;
 
-@Entity
-@Table(name = "stories")
+    @Entity
+    @Table(name = "stories")
 public class Stories {
     @Id
     @GeneratedValue
-    @SequenceGenerator(name="stories-generator")
+    @SequenceGenerator(name="prompt-id-generator")
 
+    public Long prompt_id;
     public Long story_id;
     public String title;
-    public String root_text;
-    public Long user_id;
 
     public Stories() {
     }
 
-    public Stories(String title, String root_text, Long user_id) {
+    public Stories(Long story_id, String title) {
         this.title = title;
-        this.root_text = root_text;
-        this.user_id = user_id;
+        this.story_id = story_id;
     }
 
     public String toString() {
         return this.title;
     }
 }
+
