+++ src/main/java/com/questing/littlequest/repositories/ChoiceRepository.java
@@ -3,5 +3,5 @@ package com.questing.littlequest.repositories;
 import com.questing.littlequest.models.Choices;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChoicesRepository extends JpaRepository<Choices, Long> {
+public interface ChoiceRepository extends JpaRepository<Choices, Long> {
 }
