--- src/main/java/com/questing/littlequest/repositories/NewChoiceRepository.java
@@ -1,7 +1,0 @@
-package com.questing.littlequest.repositories;
-
-import com.questing.littlequest.models.NewChoiceTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NewChoiceRepository extends JpaRepository<NewChoiceTable, Long> {
-}
