--- src/main/java/com/questing/littlequest/repositories/NewStoryRepository.java
@@ -1,6 +1,0 @@
-package com.questing.littlequest.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NewStoryRepository extends JpaRepository<NewStoryRepository, Long> {
-}
