+++ src/main/java/com/questing/littlequest/repositories/PromptRepository.java
@@ -1,0 +1,7 @@
+package com.questing.littlequest.repositories;
+
+import com.questing.littlequest.models.Prompts;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PromptRepository extends JpaRepository<Prompts, Long> {
+}
