--- src/main/java/com/questing/littlequest/repositories/StorySteveRepository.java
@@ -1,7 +1,0 @@
-package com.questing.littlequest.repositories;
-
-import com.questing.littlequest.models.StorySteve;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StorySteveRepository extends JpaRepository<StorySteve, Long> {
-}
